+++ cmd/compute/instance/instance_show.go
@@ -33,7 +33,7 @@ type InstanceShowOutput struct {
 	PublicIPAssignment    v3.PublicIPAssignment `json:"public-ip" outputLabel:"Public IP"`
 	IPAddress             string                `json:"ip_address"`
 	IPv6Address           string                `json:"ipv6_address" outputLabel:"IPv6 Address"`
-	SSHKey                string                `json:"ssh_key"`
+	SSHKeys               []string              `json:"ssh_keys"`
 	DiskSize              string                `json:"disk_size"`
 	State                 v3.InstanceState      `json:"state"`
 	Labels                map[string]string     `json:"labels"`
@@ -115,11 +115,6 @@ func (c *instanceShowCmd) CmdRun(cmd *cobra.Command, _ []string) error {
 		ipV6 = &parsed // only assign pointer if it's a valid IP
 	}
 
-	var sshKeyName *string
-	if instance.SSHKey != nil {
-		sshKeyName = &instance.SSHKey.Name
-	}
-
 	out := InstanceShowOutput{
 		AntiAffinityGroups: make([]string, 0),
 		CreationDate:       instance.CreatedAT.String(),
@@ -137,7 +132,7 @@ func (c *instanceShowCmd) CmdRun(cmd *cobra.Command, _ []string) error {
 		}(),
 		Name:            instance.Name,
 		PrivateNetworks: make([]string, 0),
-		SSHKey:          utils.DefaultString(sshKeyName, "-"),
+		SSHKeys:         make([]string, 0),
 		SecurityGroups:  make([]string, 0),
 		SecureBoot:      *instance.SecurebootEnabled,
 		Tpm:             *instance.TpmEnabled,
@@ -213,6 +208,10 @@ func (c *instanceShowCmd) CmdRun(cmd *cobra.Command, _ []string) error {
 			}
 			out.PrivateNetworks = append(out.PrivateNetworks, foundPN.Name)
 		}
+	}
+
+	for _, k := range instance.SSHKeys {
+		out.SSHKeys = append(out.SSHKeys, k.Name)
 	}
 
 	if instance.SecurityGroups != nil {